Couple improvements
