+++ lib/src/app/app_view.dart
@@ -4,13 +4,12 @@ import 'package:hive_flutter/hive_flutter.dart';
 import '../database/hive_database.dart';
 import '../edit/edit.dart';
 import '../home/home.dart';
-import 'components/components.dart';
 
 class ElPCDApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return ValueListenableBuilder(
-      valueListenable: HiveDatabase.settingsBox.listenable(),
+      valueListenable: HiveDatabase.settingsBox.listenable(keys: ['darkMode']),
       builder: (_, settingsBox, __) {
         var darkMode = settingsBox.get('darkMode', defaultValue: true);
         return MaterialApp(
@@ -25,6 +24,27 @@ class ElPCDApp extends StatelessWidget {
           },
         );
       },
+    );
+  }
+}
+
+abstract class AppTheme {
+  static ThemeData themeData(bool darkMode) {
+    return ThemeData(
+      primarySwatch: Colors.blue,
+      primaryColor: Colors.blue.shade900,
+      accentColor: Colors.orange,
+      visualDensity: VisualDensity.adaptivePlatformDensity,
+      brightness: darkMode ? Brightness.dark : Brightness.light,
+      inputDecorationTheme: InputDecorationTheme(
+        border: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(8),
+          borderSide: BorderSide(width: 2),
+        ),
+      ),
+      hoverColor: darkMode
+          ? Colors.white.withOpacity(0.15)
+          : Colors.grey.withOpacity(0.5),
     );
   }
 }
