--- lib/src/app/components/components.dart
@@ -1,1 +1,0 @@
-export 'app_theme.dart';
